docs: record nextest/strict/modernity and zero-copy rules; fix cast warning
The three operator rules land in the spec and AGENTS.md: cargo-nextest as
the test runner with strict zero-warning fmt/clippy; follow the latest
recommended practices of every ecosystem touched; and the ingest performance
posture (decode once, never clone, move ownership forward) with the
measured-gates caveat and the perf backlog (raw-bytes spool, string
interning, insert batching). Clippy's same-type-cast warning in the greptime
adapter is fixed; the workspace is back to zero warnings with seven tests
green under nextest.

+++ crates/parallax-storage/src/greptime.rs
@@ -471,7 +471,7 @@ impl TelemetryStore for GreptimeStore {
             })
             .collect();
         if agg == MetricAgg::Rate {
-            series = crate::memory::rate_from_buckets(&series, step_secs as u128 * 1_000_000_000);
+            series = crate::memory::rate_from_buckets(&series, step_secs * 1_000_000_000);
         }
         Ok(series)
     }
